+++ lib/puppet-check/ruby_parser.rb
@@ -22,8 +22,11 @@ class RubyParser
     else
       # check ruby style
       if style
+        # add rubocop rspec plugin if the file is a spec
+        rc_args.push('--plugin', 'rubocop-rspec') if file =~ /_spec\.rb$/
+
         # check RuboCop and parse warnings' JSON output
-        rubocop_warnings = Utils.capture_stdout { rubocop_cli.run(rc_args + ['--enable-pending-cops', '--plugin', 'rubocop-performance', '--plugin', 'rubocop-rspec', '--format', 'json', file]) }
+        rubocop_warnings = Utils.capture_stdout { rubocop_cli.run(rc_args + ['--enable-pending-cops', '--plugin', 'rubocop-performance', '--format', 'json', file]) }
         rubocop_offenses = JSON.parse(rubocop_warnings)['files'][0]['offenses'].map { |warning| "#{warning['location']['line']}:#{warning['location']['column']} #{warning['message']}" }
 
         # check Reek using examiner api